// database/prisma-archon/generated/zod/inputTypeSchemas/UserMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'

export const UserMaxOrderByAggregateInputSchema: z.ZodType<Prisma.UserMaxOrderByAggregateInput> = z
    .object({
        id: z.lazy(() => SortOrderSchema).optional(),
        email: z.lazy(() => SortOrderSchema).optional(),
        name: z.lazy(() => SortOrderSchema).optional(),
        firstName: z.lazy(() => SortOrderSchema).optional(),
        lastName: z.lazy(() => SortOrderSchema).optional(),
        imageUrl: z.lazy(() => SortOrderSchema).optional(),
        username: z.lazy(() => SortOrderSchema).optional(),
        profileImageUrl: z.lazy(() => SortOrderSchema).optional(),
        hasImage: z.lazy(() => SortOrderSchema).optional(),
        lastSignInAt: z.lazy(() => SortOrderSchema).optional(),
        lastActiveAt: z.lazy(() => SortOrderSchema).optional(),
        banned: z.lazy(() => SortOrderSchema).optional(),
        passwordEnabled: z.lazy(() => SortOrderSchema).optional(),
        twoFactorEnabled: z.lazy(() => SortOrderSchema).optional(),
        emailVerified: z.lazy(() => SortOrderSchema).optional(),
        leadCount: z.lazy(() => SortOrderSchema).optional(),
        plan: z.lazy(() => SortOrderSchema).optional(),
        stripeCustomerId: z.lazy(() => SortOrderSchema).optional(),
        createdAt: z.lazy(() => SortOrderSchema).optional(),
        updatedAt: z.lazy(() => SortOrderSchema).optional(),
        onboardingComplete: z.lazy(() => SortOrderSchema).optional(),
        companyName: z.lazy(() => SortOrderSchema).optional(),
        role: z.lazy(() => SortOrderSchema).optional(),
        useCase: z.lazy(() => SortOrderSchema).optional()
    })
    .strict()

export default UserMaxOrderByAggregateInputSchema